src/Jhipster/ClientApp/src/app/entities/notification-group/delete/notification-group-delete-dialog.component.spec.ts
jest.mock("@ng-bootstrap/ng-bootstrap");

import {
  ComponentFixture,
  TestBed,
  inject,
  fakeAsync,
  tick,
} from "@angular/core/testing";
import { HttpResponse } from "@angular/common/http";
import { HttpClientTestingModule } from "@angular/common/http/testing";
import { of } from "rxjs";
import { NgbActiveModal } from "@ng-bootstrap/ng-bootstrap";

import { NotificationGroupService } from "../service/notification-group.service";

import { NotificationGroupDeleteDialogComponent } from "./notification-group-delete-dialog.component";

describe("NotificationGroup Management Delete Component", () => {
  let comp: NotificationGroupDeleteDialogComponent;
  let fixture: ComponentFixture<NotificationGroupDeleteDialogComponent>;
  let service: NotificationGroupService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [NotificationGroupDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(NotificationGroupDeleteDialogComponent, "")
      .compileComponents();
    fixture = TestBed.createComponent(NotificationGroupDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(NotificationGroupService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe("confirmDelete", () => {
    it("Should call delete service on confirmDelete", inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest
          .spyOn(service, "delete")
          .mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith("deleted");
      })
    ));

    it("Should not call delete service on clear", () => {
      // GIVEN
      jest.spyOn(service, "delete");

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
